+++ packages/http-server/src/server.ts
@@ -3,7 +3,6 @@ import { createInstance, IHttpConfig, IHttpMethod, PrismHttpInstance, ProblemJso
 import { IHttpOperation } from '@stoplight/types';
 import * as fastify from 'fastify';
 import * as fastifyCors from 'fastify-cors';
-import * as formbodyParser from 'fastify-formbody';
 import { IncomingMessage, ServerResponse } from 'http';
 import { defaults } from 'lodash';
 import * as typeIs from 'type-is';
@@ -18,7 +17,7 @@ export const createServer = (operations: IHttpOperation[], opts: IPrismHttpServe
     logger: (components && components.logger) || createLogger('HTTP SERVER'),
     disableRequestLogging: true,
     modifyCoreObjects: false,
-  }).register(formbodyParser);
+  });
 
   if (opts.cors) server.register(fastifyCors);
 
@@ -30,6 +29,11 @@ export const createServer = (operations: IHttpOperation[], opts: IPrismHttpServe
         return done(e);
       }
     }
+
+    if (typeIs(req, ['application/x-www-form-urlencoded'])) {
+      return done(null, body);
+    }
+
     const error: Error & { status?: number } = new Error(`Unsupported media type.`);
     error.status = 415;
     Error.captureStackTrace(error);